Employer job authoring: create, edit, close own vacancies (task 4)
CreateVacancy reuses moderation.Service.Create via the same Minter
pattern internal/ingest/submission already uses, guarded by a
pre-check (GetJobBySourceExternalID) that refuses a URL already owned
by a different employer account before ever delegating -- closing a
real takeover gap in the existing UpsertManualJob upsert, which never
checked created_by on conflict.

UpdateVacancy/CloseVacancy use new actor-scoped queries
(UpdateEmployerJob/CloseEmployerJob), since moderation's own
Update/BySlug are scoped only to "any manually-authored job." Update
re-derives facets via job.New/jobderive.Input directly, the same
shared primitives moderation's own private derive() wraps, rather
than duplicating derivation logic or exporting that private helper.

Verified end-to-end against real Postgres: the URL-collision refusal,
the same-owner re-create reopening a closed vacancy, and the
cross-employer update/close refusal all hold with the real
moderation.Service as the Minter.

+++ internal/platform/db/queries/jobs.sql
@@ -1376,7 +1376,7 @@ SET title        = sqlc.arg(title),
                                THEN NULL ELSE jobs.similar_computed_at END,
     updated_at   = now()
 WHERE public_slug = sqlc.arg(public_slug)
-  AND created_by = sqlc.arg(actor_id)
+  AND created_by = sqlc.arg(actor_id)::bigint
   AND source = 'employer'
   AND NOT is_private
 RETURNING *;
@@ -1393,7 +1393,7 @@ WITH closed AS (
         closed_reason = 'employer_closed',
         updated_at    = now()
     WHERE jobs.public_slug = sqlc.arg(public_slug)
-      AND jobs.created_by = sqlc.arg(actor_id)
+      AND jobs.created_by = sqlc.arg(actor_id)::bigint
       AND jobs.source = 'employer'
       AND jobs.closed_at IS NULL
     RETURNING jobs.id
